requirements.txt
astropy
h5py
healpy
matplotlib>=2.0
numpy>=1.9
pandas
scipy
tqdm
emcee
dynesty
ligo.skymap
